tweeks

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: ananelli <ananelli@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/08/11 01:12:54 by ananelli          #+#    #+#             */
-/*   Updated: 2017/11/07 01:33:04 by ananelli         ###   ########.fr       */
+/*   Updated: 2017/11/07 11:10:31 by ananelli         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_strlen(const char *str)
+size_t		ft_strlen(const char *str)
 {
-	int count;
+	size_t count;
 
 	count = 0;
 	while (str[count] != '\0')
